requirements.txt
streamlit 
speech-recognition 
gtts
io 
pydub 
openai 
os 
python-dotenv
